refactor(mobile): simplify Header and migrate full height sheets to Screen
